Update radixtree to ignore values that were automatically added.
When adding a long string to radixtree if part of the string in the
middle of it did not exists in the tree before it would duplicate new
value event. Now non existent middle values are marked with null value
and nodes return parent nodes value if child has null value.